requirements.txt
Django==1.11.7
pkg-resources==0.0.0
pytz==2017.3
django-suit==0.2.25
django-import-export